refactor: move Money value class
