+++ src/test/kotlin/com/kroffle/knitting/helper/dto/MockProductData.kt
@@ -1,9 +1,9 @@
 package com.kroffle.knitting.helper.dto
 
 import com.kroffle.knitting.domain.product.entity.Product
-import com.kroffle.knitting.domain.product.value.Money
 import com.kroffle.knitting.domain.product.value.ProductItem
 import com.kroffle.knitting.domain.product.value.ProductTag
+import com.kroffle.knitting.domain.value.Money
 import com.kroffle.knitting.helper.WebTestClientHelper
 import java.time.LocalDate
 import java.time.OffsetDateTime
